+++ frontend/src/utils/__tests__/formatters.test.jsx
@@ -11,7 +11,6 @@ import {
   dateFormatter,
   phoneNumberFormatter,
   formatNumberWithCommas,
-  isArrayEmpty,
   spacesFormatter,
   timezoneFormatter,
   convertObjectKeys
@@ -375,19 +374,5 @@ describe('formatNumberWithoutCommas', () => {
 
   it('should return undefined for invalid numbers', () => {
     expect(formatNumberWithoutCommas('abc')).toEqual(undefined)
-  })
-})
-
-describe('isArrayEmpty', () => {
-  it('should return true for empty arrays', () => {
-    expect(isArrayEmpty([])).toEqual(true)
-  })
-
-  it('should return false for non-empty arrays', () => {
-    expect(isArrayEmpty([1, 2, 3])).toEqual(false)
-  })
-
-  it('should return null for non-array inputs', () => {
-    expect(isArrayEmpty({})).toEqual(null)
   })
 })
