+++ frontend/src/views/FuelSupplies/AddEditFuelSupplies.jsx
@@ -7,19 +7,16 @@ import {
   useGetFuelSupplies,
   useSaveFuelSupply
 } from '@/hooks/useFuelSupply'
-import { isArrayEmpty, cleanEmptyStringValues } from '@/utils/formatters'
+import { cleanEmptyStringValues } from '@/utils/formatters'
 import BCTypography from '@/components/BCTypography'
 import Grid2 from '@mui/material/Unstable_Grid2/Grid2'
 import { useCallback, useEffect, useMemo, useRef, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import { useLocation, useNavigate, useParams } from 'react-router-dom'
 import { v4 as uuid } from 'uuid'
-import {
-  defaultColDef,
-  fuelSupplyColDefs,
-  PROVISION_APPROVED_FUEL_CODE
-} from './_schema'
+import { defaultColDef, fuelSupplyColDefs } from './_schema'
 import { handleScheduleDelete, handleScheduleSave } from '@/utils/schedules.js'
+import { isArrayEmpty } from '@/utils/array.js'
 
 export const AddEditFuelSupplies = () => {
   const [rowData, setRowData] = useState([])
